プルリクで指摘受けた個所をint型に戻し、float型で引数で渡す個所はstaticキャストしました

+++ NinjaResidence/FallRock.cpp
@@ -52,14 +52,14 @@ void FallRock::Render(int MapScrollY, int MapScrollX, MapDataState MapDataRevers
 		return;
 	}
 
-	m_GimmickVertex[0].x = m_GimmickPosX + MapScrollX;
-	m_GimmickVertex[0].y = m_GimmickPosY + MapScrollY + m_QuantityOfMovement;
-	m_GimmickVertex[1].x = m_GimmickPosX + (CELL_SIZE * 2.f) + MapScrollX;
-	m_GimmickVertex[1].y = m_GimmickPosY + MapScrollY + m_QuantityOfMovement;
-	m_GimmickVertex[2].x = m_GimmickPosX + (CELL_SIZE * 2.f) + MapScrollX;
-	m_GimmickVertex[2].y = m_GimmickPosY + (CELL_SIZE * 2.f) + MapScrollY + m_QuantityOfMovement;
-	m_GimmickVertex[3].x = m_GimmickPosX + MapScrollX;
-	m_GimmickVertex[3].y = m_GimmickPosY + (CELL_SIZE * 2.f) + MapScrollY + m_QuantityOfMovement;
+	m_GimmickVertex[0].x = static_cast<float>(m_GimmickPosX + MapScrollX);
+	m_GimmickVertex[0].y = static_cast<float>(m_GimmickPosY + MapScrollY + m_QuantityOfMovement);
+	m_GimmickVertex[1].x = static_cast<float>(m_GimmickPosX + (CELL_SIZE * 2.f) + MapScrollX);
+	m_GimmickVertex[1].y = static_cast<float>(m_GimmickPosY + MapScrollY + m_QuantityOfMovement);
+	m_GimmickVertex[2].x = static_cast<float>(m_GimmickPosX + (CELL_SIZE * 2.f) + MapScrollX);
+	m_GimmickVertex[2].y = static_cast<float>(m_GimmickPosY + (CELL_SIZE * 2.f) + MapScrollY + m_QuantityOfMovement);
+	m_GimmickVertex[3].x = static_cast<float>(m_GimmickPosX + MapScrollX);
+	m_GimmickVertex[3].y = static_cast<float>(m_GimmickPosY + (CELL_SIZE * 2.f) + MapScrollY + m_QuantityOfMovement);
 
 	m_GimmickVertex[0].tu = BLOCK_INTEGRATION_WIDTH * 4.f;
 	m_GimmickVertex[1].tu = BLOCK_INTEGRATION_WIDTH * 6.f;
